pyproject.toml
[project]
name = "svaha-mini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "aio-pika>=9.5.4",
    "aioboto3>=13.3.0",
    "aioredis>=2.0.1",
    "fastapi>=0.115.6",
    "orjson>=3.10.14",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "starlette==0.45.2",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "ulid-py>=1.1.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.0",
]

[tool.black]
skip-string-normalization = false
line-length = 120
#extend-ignore = "E203"  # todo It's not working. Fix it by adding E203 and preview=true to the Ruff

[tool.ruff]
line-length = 120
preview = true

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
ignore = [
    "ANN101", # Missing type annotation for self in a class method
    "ANN102", # Missing type annotation for cls in a classmethod
    "C408", # Unnecessary `dict` call
    "PLR2004", # Magic value used in comparison
    "RET504", # Unnecessary variable assignment before `return` statement
    "RUF002", # Docstring contains ambiguous unicode character
    "RUF003", # Comment contains ambiguous unicode character
    "UP006", # Use `list` instead of `List` for type annotations
    "UP007"  # Use `X | Y` for type annotations
]
select = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = false
force-single-line = true
known-first-party = [""]
order-by-type = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.flake8-quotes]
#docstring-quotes = "single"
inline-quotes = "single"
#multiline-quotes = "single"

# Test
# black --diff --color .
# ruff check --diff . | colordiff

# Fix
# black .
# ruff check --fix .


# Test
# rye add --dev pytest
# rye test

# Run
# uvicorn --reload --host $HOST --port $PORT --log-level $LOG_LEVEL --log-config log_config.json
